Rename frscrcmd_battle_hall
