+++ platinum.us/main.lsf
@@ -1543,7 +1543,7 @@ Overlay overlay104
 	Object main.nef.p/src_overlay104_ov104_02231F74.c.o
 	Object main.nef.p/src_overlay104_ov104_022332B4.c.o
 	Object main.nef.p/src_overlay104_ov104_022339B4.c.o
-	Object main.nef.p/src_overlay104_ov104_02234838.c.o
+	Object main.nef.p/src_overlay104_frscrcmd_battle_hall.c.o
 	Object main.nef.p/src_overlay104_ov104_02234DB4.c.o
 	Object main.nef.p/src_overlay104_ov104_022358E8.c.o
 	Object main.nef.p/src_overlay104_ov104_022361B4.c.o
