+++ src/overlay104/frscrcmd.c
@@ -24,11 +24,11 @@
 #include "overlay104/frontier_particle_system.h"
 #include "overlay104/frontier_script_context.h"
 #include "overlay104/frontier_script_manager.h"
+#include "overlay104/frscrcmd_battle_hall.h"
 #include "overlay104/frscrcmd_sound.h"
 #include "overlay104/ov104_0222DCE0.h"
 #include "overlay104/ov104_02231F74.h"
 #include "overlay104/ov104_022332B4.h"
-#include "overlay104/ov104_02234838.h"
 #include "overlay104/ov104_022358E8.h"
 #include "overlay104/ov104_02237378.h"
 #include "overlay104/ov104_02239130.h"
